Refactor `BaseSettingsPage` to `BasePage`
