+++ src/main/java/io/redstudioragnarok/fbp/gui/menu/Page3.java
@@ -1,13 +1,14 @@
 package io.redstudioragnarok.fbp.gui.menu;
 
 import io.redstudioragnarok.fbp.FBP;
+import io.redstudioragnarok.fbp.gui.BasePage;
 import io.redstudioragnarok.fbp.handlers.ConfigHandler;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.resources.I18n;
 
 import static io.redstudioragnarok.fbp.gui.FBPGuiButton.ButtonSize.large;
 
-public class Page3 extends BaseSettingsPage {
+public class Page3 extends BasePage {
 
 	GuiButton fancyFlame, fancySmoke, fancyWeather, dynamicWeather, waterPhysics, restOnFloor;
